Split out a few parts of the code into files
